# laravel11/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SyncBridge API",
        "description": "Production-grade API documentation for SyncBridge integration endpoints.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080",
            "description": "Primary API server"
        }
    ],
    "paths": {
...
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List orders",
                "description": "Returns paginated order records.",
                "operationId": "ordersIndex",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 200,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiSuccessBase"
                                        },
                                        {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 80
                                                },
                                                "page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "rows": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Order"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{orderId}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order details",
                "description": "Returns order details and order items.",
                "operationId": "ordersShow",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiSuccessBase"
                                        },
                                        {
                                            "properties": {
                                                "order": {
                                                    "$ref": "#/components/schemas/Order"
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/OrderItem"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiSuccessBase": {
                "required": [
                    "ok",
                    "message"
                ],
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    }
                },
                "type": "object"
            },
            "ApiError": {
                "required": [
                    "ok",
                    "error",
                    "errors"
                ],
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "string",
                        "example": "Invalid payload"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "schema": "OpenApiSpec",
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "The field is required."
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "Order": {
                "required": [
                    "id",
                    "sync_status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 14
                    },
                    "ay_order_id": {
                        "type": "string",
                        "example": "AY-ORDER-123",
                        "nullable": true
                    },
                    "ps_order_id": {
                        "type": "integer",
                        "example": 345,
                        "nullable": true
                    },
                    "sync_status": {
                        "type": "string",
                        "example": "synced"
                    },
                    "ay_status": {
                        "type": "string",
                        "example": "approved",
                        "nullable": true
                    },
                    "total_paid": {
                        "type": "number",
                        "format": "float",
                        "example": 129.49
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 55
                    },
                    "order_id": {
                        "type": "integer",
                        "example": 14
                    },
                    "ay_order_item_id": {
                        "type": "string",
                        "example": "item-001",
                        "nullable": true
                    },
                    "sku": {
                        "type": "string",
                        "example": "SKU-TEE-M",
                        "nullable": true
                    },
                    "ean13": {
                        "type": "string",
                        "example": "4006381333931",
                        "nullable": true
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 100,
                        "nullable": true
                    },
                    "combo_id": {
                        "type": "integer",
                        "example": 200,
                        "nullable": true
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "unit_price": {
                        "type": "number",
                        "format": "float",
                        "example": 19.99
                    },
                    "discount_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "item_status": {
                        "type": "string",
                        "example": "shipped",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Paste SYNCBRIDGE_API_TOKEN as Bearer token.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            },
            "apiTokenHeader": {
                "type": "apiKey",
                "description": "Alternative auth header for sync-protected endpoints.",
                "name": "X-Api-Token",
                "in": "header"
            }
        }
    },
...
}